Parser improvements, started writing docs

+++ src/number.cpp
@@ -11,7 +11,7 @@ number::number(const double &n)
     else
         v = n;
 };
-number::number(const size_t &n) : v(static_cast<int>(n)){};
+number::number(const unsigned int &n) : v(static_cast<int>(n)){};
 number::number(const std::string &n)
 {
     if (n.find('.') == std::string::npos)
